Revert "refactor: 디렉토리 구조 변경"
This reverts commit 69c344aa18984a535d7ffc89c94240ff12ab5797.
