+++ app/src/main/java/com/example/mvc_project/model/dto/PlotsDTO.kt
@@ -1,4 +1,4 @@
-package com.example.mvc_project.data.dto
+package com.example.mvc_project.model.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
