+++ app/src/main/java/com/example/mvc_project/model/network/RetrofitInstance.kt
@@ -1,4 +1,4 @@
-package com.example.mvc_project.data.api
+package com.example.mvc_project.model.network
 
 import com.example.mvc_project.BuildConfig
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
