+++ app/src/main/java/com/example/mvc_project/view/MovieAdapter.kt
@@ -1,5 +1,6 @@
-package com.example.mvc_project.ui.main
+package com.example.mvc_project.view
 
+import com.example.mvc_project.R
 import android.annotation.SuppressLint
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -7,9 +8,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.example.mvc_project.R
 import com.example.mvc_project.databinding.ItemBinding
-import com.example.mvc_project.domain.model.MovieUiState
 
 class MovieAdapter(
     private val onMovieItemClick: (MovieUiState) -> Unit
