+++ app/src/main/java/com/example/mvc_project/view/MovieDetailUiState.kt
@@ -1,4 +1,4 @@
-package com.example.mvc_project.domain.model
+package com.example.mvc_project.view
 
 data class MovieDetailUiState(
     val moviePoster: String = "",
